Refactor vendor locking to use explicit lockfile
Lock in `vendor_lock.json` and keep track of all content hashes there.
That means that this repo has tracking for updates and will be able to
check for updates by looking for an update to the vendor lock.

This flow should work out better for enabling a check to see if there
were changes at the time of a potential release, without having to
check the last release package from pypi or a similar source (e.g.
github release artifacts).